# dpki_api/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Distributed PKI management API"
  version: "1.0"
  title: "Distributed PKI API"
host: "virtserver.swaggerhub.com"
basePath: "/ndid/dpki/0.1"
schemes:
- "https"
paths:
  /dpki/node/create:
    post:
      summary: "create node in the dpki module"
      description: "NDID ONLY - used by NDID to add new organization to the network.\
        \  When an organization register to host NDID node, it must present the node_data\
        \ to NDID (node_id, node_name, node_key, node_master_key)  node_id must be\
        \ unique."
      operationId: "dpkiNodeCreatePOST"
      parameters:
      - in: "body"
        name: "node_data"
        required: true
        schema:
          $ref: "#/definitions/node_data"
      responses:
        200:
          description: "successful"
        400:
          description: "error, such as duplicate node_id"
      x-swagger-router-controller: "Default"
  /dpki/node/update:
    post:
      summary: "update node name"
      description: "Update node_name, node_key or node_master_key"
      operationId: "dpkiNodeUpdatePOST"
      parameters:
      - in: "body"
        name: "node_data"
        required: true
        schema:
          $ref: "#/definitions/node_data"
      responses:
        200:
          description: "successful"
        400:
          description: "error: unauthorized"
      x-swagger-router-controller: "Default"
  /dpki/node/register_callback:
    post:
      summary: "Register a callback location for NDID node to call."
      description: "Register a callback location for NDID DPKI node to call to get\
        \ things signed.  Required if want to use HSM or other methods of key signing\
        \ and storage other than default (store at node app data)"
      operationId: "dpkiNodeRegister_callbackPOST"
      parameters:
      - in: "body"
        name: "url"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Registration Successful"
      x-swagger-router-controller: "Default"
  /dpki/node/register_callback_master:
    post:
      summary: "Register a callback location for NDID node to call when it needs master\
        \ key."
      description: "Register a callback location for NDID DPKI node to call to get\
        \ things signed by master key.  Required if want to use HSM or other methods\
        \ of key signing and storage other than default (store at node app data)"
      operationId: "dpkiNodeRegister_callback_masterPOST"
      parameters:
      - in: "body"
        name: "url"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Registration Successful"
      x-swagger-router-controller: "Default"
definitions:
  node_data:
    type: "object"
    required:
    - "node_id"
    - "node_key"
    - "node_master_key"
    - "node_name"
    properties:
      node_id:
        type: "string"
      node_name:
        type: "string"
        description: "Human readable name of the node, publicly accessible.  Possibly\
          \ organization name."
      node_key:
        type: "string"
        description: "Node's Public Key"
      node_key_type:
        type: "string"
        description: "Type of Key (For example: RSA1024, ECC ED25519)"
      node_key_method:
        type: "string"
        description: "Name of Signature algorithm for this key"
      node_master_key:
        type: "string"
        description: "Node's Master Key, this is a public key that is used to manage\
          \ node's data and update node's public key.  Master key can only be updated\
          \ using previous master key, whereas node's public key can be updated using\
          \ previous public key or master key"
      node_master_key_type:
        type: "string"
        description: "Type of Key (For example: RSA1024, ECC ED25519)"
      node_master_key_method:
        type: "string"
        description: "Name of Signature algorithm for this key"